systemd: Implement crypto policies support
This implements the ability to set and show crypto policies from within
the cockpit system overview. The current implementation is limited to
three hardcoded policies and `update-security-policies` does not have an
option to list all existing policies. When a crypto policy is applied
but a user chooses to not reboot a notification is shown in the health
card overview.

Fixes #8519