test: Properly clean up old admin user in check-system-realm
When removing the old local "admin" user account, remove its home
directory as well. Otherwise the new identity domain manged "admin" user
will not be able to access it (due to different UID). Instead, a new
home directory will be created on demand on first login.

Use the same approach in TestKerberos, as that's cleaner as the chown
approach. However, keep it in testQualifiedUsers(), as the test switches
back and forth between both admins.