test: run-tests dynamic serial test scheduling
Computing the batches in advance by just splitting the list into equal
parts does not work well: It does not take retries and wildy different
durations into account, and thus often created a wild imbalance beween
the total runtime on the serial runner machines:

    4592s on [host], 118 serial tests: 0: 4589s, 1: 2078s

In many cases this even led to a serial test batch dominate the total
test time.

Move to dynamically assignment of the next serial test from the queue to
a free serial machine to avoid that. Drop the complicated "batch_tests"
map and instead move the per-machine test time into GlobalMachine, and
the "to which machine is a test assigned" bookkeeping into the Test
class.

Also kill the global machines as soon as they are not needed any more,
to free up resources for the parallel tests.

Implement `--list` in a separate list_tests() function. detect_tests()
already does a fine job of determining the test names, we don't have to
re-run all the check-* programs again. Instead just print their names,
and sort them (independently of $IMAGE).  This makes `--list` a lot
faster, and also avoids lots of special cases in run(). Also drop the
"1..N" TAP header from the list output, as this is very much *not* TAP.
Adjust check-testlib accordingly.