test: Fix TestSession
Actually stop the Cockpit (and thus logind) session with a logout, as
the comment suggests, and also check that the logind session is going
away afterwards. It could otherwise happen that the ssh-agent keeps
running and the session remains in state "closing" forever.

This broke "affected test" retries a lot. But even if that didn't
happen, having to wait for the bridge timeout makes the test take
unnecessarily long.