cockpitwebresponse: add gzip support
Add a new helper: cockpit_web_response_file_or_gz().

If the requested file isn't found, then try looking for it with '.gz'
added to the end.  If that works, and the client accepts gzip, send it
as-is.  Otherwise, decompress it before sending.

Use this from the test-server to allow serving compressed resources from
in-tree.