test: Increase session cleanup hammer size on Fedora 34
Fedora 34's logind/systemd often leaves "closing" state logind sessions
around with *zero* processes. Explicitly kill the scope units to clean
up the session.

This is an utter hack, but as long as it does not happen any more on
newer versions, it will go away soon (as we just stopped updating Fedora
34 anyway).