test: Allow exceptions while waiting
ph_wait now allows all exceptions and keeps trying.  If the last try
has an exception, it will be reported.  This means that we don't lose
any exceptions that happen consistently, but we always have to wait
for the timeout.

Browser.wait_js_cond retries when the execution context is not found
or disappears in the middle.  To support this, getFrameExecId never
raises an exception itself but returns -1 so that Runtime.evaluate
raises an error that causes Browser.wait_js_cond to retry.

Together these changes mean that calls to functions like
Browser.wait_in_text can now survive a page reload in their middle.

Consequently, all calls to expect_load and expect_load_frame have been
removed.

Specifically, this fixes a race in the TestKeys.testAuthorizedKeys
test case.  That test case experiences a page reload that is triggered
asynchronously and can therefore not be reliably waited for with
expect_load_frame.  (The call to enter_page was essentially a no-op.)