ws: fix a bug in cockpit_auth_login_async()
cockpit_auth_login_async() is a complex operation spanning many
functions, and it can succeed and fail under a wide range of different
circumstances.

One thing that all of these circumstances have in common is that, at the
time the result is dispatched back to the caller, the GSimpleAsyncResult
object is unreffed and cleared from the session struct.

There is one case where we miss this, though: in case we are calling
_login_async() to reply to an existing conversation and for some reason
reply_authorize_challenge() fails, then we dispatch an error, but don't
clear the GAsyncResult from the session.

This is the subtle issue that has been preventing us from porting to
GTask for so long.  With this issue addressed, the GTask port is (for
the most part) straightforward.