test: Split off TestMenu.testSessionTimeout
testBasic is a grab-bag of unrelated small tests. The session timeout
handling is a large and bulky test however, and one does not always want
to suffer through that when e.g. investigating aXe issues or the
override handling.

Split it into two tests. This also makes the timeout test less dependent
on what happens before starting the idle timeout.

Make the tests nondestructive.