test: Make load-hog.service heavier
On Fedora 36, the current command is not enough any more to create a
high enough load. It looks like /dev/urandom became a lot faster with
the current kernel?

So increase the work that it needs to do by 10x. It gets killed as soon
as the load is > 15, so that doesn't increase test time on the other
OSes.