build: fix webpack build for out-of-tree
We need to invoke webpack from the srcdir, so change to there first.

Take the chance to simplify this rule: the unnecessary use of variables
makes it difficult to understand, and what's going on here is actually
pretty simple.

Adjust some other rules in pkg/build to make it clear which files are
expected to be in $(srcdir).  make can't always figure it out.