test: Make testNoAdminGroup nondestructive and move it to TestLogin
This did not fit well into `TestShutdownRestart`, as it does not need or
make use of the second provisioned machine. It's also not necessary to
actually reboot the machine, we just want to check that the user is
privileged enough to do so. Doing a scheduled reboot and cancelling it
is just as good, and allows us to make the test nondestructive.