systemd: check if command exists in timer creation dialog
As we currently have no timer deletion/edit option it's useful to know
if the created timer is will actually run succesfully. Checking if the
command exists is an easy check to add, the commit does not change field
to a fileinput as a user might want to add arguments to it's command.