test: Clean up and factorize chrony hack in check-system-realms
This is not a distro bug, at least not in the way we thought and
reported. Our images explicitly install systemd-timesync (for testing
the Overview page), which conflicts with the chrony package. So our
approach of just mocking chrony.service was justified. Update its
description, and factorize it into a helper function.

Restrict the "fake chronyd.service" hack to ubuntu-2004, this bug is
fixed in later versions.