test: Fix teslib test
This test was not doing at all what it was supposed to do.
The main issue was that it created `test-shell-testlib` and not
`check-shell-testlib` thus was not recognized at all.
The other issue was that the replace did nothing, as `NoTest` is already
changed a few lines above. And lastly, even if it would work, it would
still fail as there were conflicting test names in this and the original
file.